MAX77696: Changed driver load order to device
MAX77696 driver was previously loaded as arch_initcall
which caused it to load be loaded before the i2c
subsystem on imx6sl was fully loaded, resulting in
the MAX77696 driver silently giving up.

Moved MAX77696 down the initcall order from
arch_initcall to device_initcall to resolve the issue.

Signed-off-by: Jacob Postman <jacob@revolution-robotics.com>